AVQA/requirements.txt
certifi==2022.9.24
einops==0.4.1
ffmpeg==1.4
gpuinfo==1.0.0a7
ipdb==0.13.9
munch==2.5.0
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==9.3.0
psutil==5.9.2
timm==0.6.7
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
